2006-03-27  Christian Stimming <stimming@tuhh.de>
        * src/parser/xsd_write.c, src/net2/nl_http.c: Add cast to
        isdigit() et al functions; needed for platforms where "char" has a
        different signedness than on most platforms. Notably
        sparc-solaris.

        * configure.ac: Change gcc version number checking so that it
        works with 2.95.x as well.


git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1024 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/net2/nl_hbci.c
@@ -357,7 +357,7 @@ GWEN_NETLAYER_RESULT GWEN_NetLayerHbci__ReadWork(GWEN_NETLAYER *nl) {
     /* read size */
     p++;
     bsize=0;
-    while(*p && isdigit(*p)) {
+    while(*p && isdigit((int)*p)) {
       bsize*=10;
       bsize+=*p-'0';
       p++;
